More linter changes
