+++ src/util/enum.ts
@@ -16,3 +16,11 @@ export enum RecordCompleteness {
 export enum HttpMethod {
   PUT = 'put',
 }
+
+export enum Status {
+  ARCHIVED = "archived",
+  CURRENT = "current",
+  PROVISIONAL = "provisional",
+  PROVISIONAL_OVER_CURRENT = "provisional_over_current",
+  ALL = "all"
+}
