--- src/util/enums.ts
@@ -1,7 +1,0 @@
-export enum Status {
-    ARCHIVED = "archived",
-    CURRENT = "current",
-    PROVISIONAL = "provisional",
-    PROVISIONAL_OVER_CURRENT = "provisional_over_current",
-    ALL = "all"
-}
